CDAP-20113 - Offload plugin execution to Task Workers in security enabled configurations
