+++ delta-app/src/main/java/io/cdap/delta/app/service/AssessmentHandler.java
@@ -20,6 +20,9 @@ import com.google.gson.JsonSyntaxException;
 import io.cdap.cdap.api.plugin.PluginConfigurer;
 import io.cdap.cdap.api.service.http.HttpServiceRequest;
 import io.cdap.cdap.api.service.http.HttpServiceResponder;
+import io.cdap.cdap.api.service.worker.RemoteExecutionException;
+import io.cdap.cdap.api.service.worker.RemoteTaskException;
+import io.cdap.cdap.api.service.worker.RunnableTaskRequest;
 import io.cdap.delta.api.DeltaPipelineId;
 import io.cdap.delta.api.ReplicationError;
 import io.cdap.delta.api.assessment.PipelineAssessment;
@@ -29,6 +32,7 @@ import io.cdap.delta.app.DefaultConfigurer;
 import io.cdap.delta.app.DeltaWorkerId;
 import io.cdap.delta.app.PipelineStateService;
 import io.cdap.delta.app.service.common.AbstractAssessorHandler;
+import io.cdap.delta.app.service.remote.*;
 import io.cdap.delta.proto.DBTable;
 import io.cdap.delta.proto.DeltaConfig;
 import io.cdap.delta.proto.DraftRequest;
@@ -38,9 +42,9 @@ import io.cdap.delta.proto.StateRequest;
 import io.cdap.delta.proto.TableAssessmentRequest;
 import io.cdap.delta.proto.TableAssessmentResponse;
 import io.cdap.delta.proto.TableReplicationState;
-import io.cdap.delta.store.Draft;
-import io.cdap.delta.store.DraftId;
-import io.cdap.delta.store.StateStore;
+import io.cdap.delta.store.*;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
 import java.net.HttpURLConnection;
 import java.nio.charset.StandardCharsets;
@@ -48,6 +52,7 @@ import java.util.Collection;
 import java.util.HashSet;
 import java.util.List;
 import java.util.Set;
+import javax.annotation.Nullable;
 import javax.ws.rs.DELETE;
 import javax.ws.rs.GET;
 import javax.ws.rs.POST;
@@ -59,6 +64,7 @@ import javax.ws.rs.PathParam;
  * Handler for storing drafts and performing assessments.
  */
 public class AssessmentHandler extends AbstractAssessorHandler {
+  private static final Logger LOG = LoggerFactory.getLogger(AssessmentHandler.class.getName());
 
   @GET
   @Path("v1/contexts/{context}/drafts")
@@ -91,9 +97,11 @@ public class AssessmentHandler extends AbstractAssessorHandler {
       try {
         draft = GSON.fromJson(StandardCharsets.UTF_8.decode(request.getContent()).toString(), DraftRequest.class);
       } catch (JsonSyntaxException e) {
+        LOG.error("Error in decoding request body:", e);
         responder.sendError(HttpURLConnection.HTTP_BAD_REQUEST, "Unable to decode request body: " + e.getMessage());
         return;
       } catch (IllegalArgumentException e) {
+        LOG.error("Invalid request: ", e);
         responder.sendError(HttpURLConnection.HTTP_BAD_REQUEST, "Invalid config: " + e.getMessage());
         return;
       }
@@ -121,9 +129,17 @@ public class AssessmentHandler extends AbstractAssessorHandler {
                               @PathParam("draft") String draftName) {
     respond(namespaceName, responder, (namespace) -> {
       DraftId draftId = new DraftId(namespace, draftName);
-      PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
-      TableList tableList = getDraftService().listDraftTables(draftId, new DefaultConfigurer(pluginConfigurer));
-      responder.sendString(GSON.toJson(tableList));
+      Draft draft = getDraftService().getDraft(draftId);
+
+      LOG.warn("isRemoteTaskEnabled: "+getContext().isRemoteTaskEnabled());
+      if (getContext().isRemoteTaskEnabled()) {
+        executeRemotely(namespace, null, draft.getConfig(), RemoteListTablesTask.class, responder);
+      } else {
+        PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
+        TableList tableList = getDraftService().listDraftTables(draftId, draft,
+                                                                new DefaultConfigurer(pluginConfigurer));
+        responder.sendString(GSON.toJson(tableList));
+      }
     });
   }
 
@@ -133,24 +149,32 @@ public class AssessmentHandler extends AbstractAssessorHandler {
                             @PathParam("context") String namespaceName,
                             @PathParam("draft") String draftName) {
     respond(namespaceName, responder, (namespace) -> {
+      String requestContent = StandardCharsets.UTF_8.decode(request.getContent()).toString();
       DBTable dbTable;
       try {
-        dbTable = GSON.fromJson(StandardCharsets.UTF_8.decode(request.getContent()).toString(), DBTable.class);
+        dbTable = GSON.fromJson(requestContent, DBTable.class);
         dbTable.validate();
       } catch (JsonSyntaxException e) {
+        LOG.error("Error in decoding request body:", e);
         responder.sendError(HttpURLConnection.HTTP_BAD_REQUEST, "Unable to decode request body: " + e.getMessage());
         return;
       } catch (IllegalArgumentException e) {
+        LOG.error("Invalid request: ", e);
         responder.sendError(HttpURLConnection.HTTP_BAD_REQUEST, "Invalid request: " + e.getMessage());
         return;
       }
 
       DraftId draftId = new DraftId(namespace, draftName);
-      PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
-      TableDetail tableDetail = getDraftService()
-        .describeDraftTable(draftId, new DefaultConfigurer(pluginConfigurer), dbTable.getDatabase(),
-          dbTable.getSchema(), dbTable.getTable());
-      responder.sendString(GSON.toJson(tableDetail));
+      Draft draft = getDraftService().getDraft(draftId);
+
+      if (getContext().isRemoteTaskEnabled()) {
+        executeRemotely(namespace, requestContent, draft.getConfig(), RemoteDescribeTableTask.class, responder);
+      } else {
+        PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
+        TableDetail tableDetail = getDraftService()
+                .describeDraftTable(draftId, new DefaultConfigurer(pluginConfigurer), dbTable);
+        responder.sendString(GSON.toJson(tableDetail));
+      }
     });
   }
 
@@ -161,10 +185,18 @@ public class AssessmentHandler extends AbstractAssessorHandler {
                           @PathParam("draft") String draftName) {
     respond(namespaceName, responder, ((namespace) -> {
       DraftId draftId = new DraftId(namespace, draftName);
-      PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
-      PipelineAssessment assessment =
-        getDraftService().assessPipeline(draftId, new DefaultConfigurer(pluginConfigurer));
-      responder.sendString(GSON.toJson(assessment));
+      Draft draft = getDraftService().getDraft(draftId);
+
+      draft.getConfig().validatePipeline();
+
+      if (getContext().isRemoteTaskEnabled()) {
+        executeRemotely(namespace, null, draft.getConfig(), RemoteAssessPipelineTask.class, responder);
+      } else {
+        PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
+        PipelineAssessment assessment =
+                getDraftService().assessPipeline(namespace, draft, new DefaultConfigurer(pluginConfigurer));
+        responder.sendString(GSON.toJson(assessment));
+      }
     }));
   }
 
@@ -177,15 +209,21 @@ public class AssessmentHandler extends AbstractAssessorHandler {
       try {
         deltaConfig = GSON.fromJson(StandardCharsets.UTF_8.decode(request.getContent()).toString(), DeltaConfig.class);
       } catch (JsonSyntaxException e) {
+        LOG.error("Error in decoding request body:", e);
         responder.sendError(HttpURLConnection.HTTP_BAD_REQUEST, "Unable to decode request body , " +
           "doesn't conform to deltaConfig: " + e.getMessage());
         return;
       }
+      deltaConfig.validatePipeline();
 
-      PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
-      PipelineAssessment assessment =
-        getDraftService().assessPipeline(namespace, deltaConfig, new DefaultConfigurer(pluginConfigurer));
-      responder.sendString(GSON.toJson(assessment));
+      if (getContext().isRemoteTaskEnabled()) {
+        executeRemotely(namespace, null, deltaConfig, RemoteAssessPipelineTask.class, responder);
+      } else {
+        PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
+        PipelineAssessment assessment =
+                getDraftService().assessPipeline(namespace, deltaConfig, new DefaultConfigurer(pluginConfigurer));
+        responder.sendString(GSON.toJson(assessment));
+      }
     }));
   }
 
@@ -196,22 +234,31 @@ public class AssessmentHandler extends AbstractAssessorHandler {
                           @PathParam("draft") String draftName) {
     respond(namespaceName, responder, ((namespace) -> {
       DraftId draftId = new DraftId(namespace, draftName);
+      String requestContent = StandardCharsets.UTF_8.decode(request.getContent()).toString();
       DBTable dbTable;
       try {
-        dbTable = GSON.fromJson(StandardCharsets.UTF_8.decode(request.getContent()).toString(), DBTable.class);
+        dbTable = GSON.fromJson(requestContent, DBTable.class);
         dbTable.validate();
       } catch (JsonSyntaxException e) {
+        LOG.error("Error in decoding request body:", e);
         responder.sendError(HttpURLConnection.HTTP_BAD_REQUEST, "Unable to decode request body: " + e.getMessage());
         return;
       } catch (IllegalArgumentException e) {
+        LOG.error("Invalid request: ", e);
         responder.sendError(HttpURLConnection.HTTP_BAD_REQUEST, "Invalid request: " + e.getMessage());
         return;
       }
 
-      PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
-      TableAssessmentResponse assessment = getDraftService()
-        .assessTable(draftId, new DefaultConfigurer(pluginConfigurer), dbTable);
-      responder.sendString(GSON.toJson(assessment));
+      Draft draft = getDraftService().getDraft(draftId);
+
+      if (getContext().isRemoteTaskEnabled()) {
+        executeRemotely(namespace, requestContent, draft.getConfig(), RemoteAssessTableTask.class, responder);
+      } else {
+        PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
+        TableAssessmentResponse assessment = getDraftService()
+          .assessTable(namespace, draft, new DefaultConfigurer(pluginConfigurer), dbTable);
+        responder.sendString(GSON.toJson(assessment));
+      }
     }));
   }
 
@@ -225,16 +272,24 @@ public class AssessmentHandler extends AbstractAssessorHandler {
         tableAssessmentRequest = GSON.fromJson(StandardCharsets.UTF_8.decode(request.getContent()).toString(),
                                                TableAssessmentRequest.class);
       } catch (JsonSyntaxException e) {
+        LOG.error("Error in decoding request body:", e);
         responder.sendError(HttpURLConnection.HTTP_BAD_REQUEST, "Unable to decode request body , " +
           "doesn't conform to TableAssessmentRequest: " + e.getMessage());
         return;
       }
 
-      PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
-      TableAssessmentResponse assessment = getDraftService()
-        .assessTable(namespace, tableAssessmentRequest.getDeltaConfig(), new DefaultConfigurer(pluginConfigurer),
-                     tableAssessmentRequest.getDBTable());
-      responder.sendString(GSON.toJson(assessment));
+      DeltaConfig deltaConfig = tableAssessmentRequest.getDeltaConfig();
+      DBTable dbTable = tableAssessmentRequest.getDBTable();
+
+      if (getContext().isRemoteTaskEnabled()) {
+        executeRemotely(namespace, GSON.toJson(dbTable), deltaConfig, RemoteAssessTableTask.class, responder);
+      } else {
+        PluginConfigurer pluginConfigurer = getContext().createPluginConfigurer(namespaceName);
+        TableAssessmentResponse assessment = getDraftService()
+                .assessTable(namespace, deltaConfig, new DefaultConfigurer(pluginConfigurer),
+                        tableAssessmentRequest.getDBTable());
+        responder.sendString(GSON.toJson(assessment));
+      }
     }));
   }
 
@@ -256,9 +311,11 @@ public class AssessmentHandler extends AbstractAssessorHandler {
                                      StateRequest.class);
         stateRequest.validate();
       } catch (JsonSyntaxException e) {
+        LOG.error("Error in decoding request body:", e);
         responder.sendError(HttpURLConnection.HTTP_BAD_REQUEST, "Unable to decode request body: " + e.getMessage());
         return;
       } catch (IllegalArgumentException e) {
+        LOG.error("Invalid request: ", e);
         responder.sendError(HttpURLConnection.HTTP_BAD_REQUEST, "Invalid request: " + e.getMessage());
         return;
       }
@@ -296,5 +353,46 @@ public class AssessmentHandler extends AbstractAssessorHandler {
       }
       responder.sendString(GSON.toJson(new PipelineReplicationState(sourceState, tableStates, sourceError)));
     }));
+  }
+
+  /**
+   * Common method for all remote executions.
+   * Remote request is created, executed and response is added to {@link HttpServiceResponder}
+   * @param namespace namespace string
+   * @param request Serialized request string
+   * @param config {@link Draft} details if present
+   * @param remoteExecutionTaskClass Remote execution task class
+   * @param responder {@link HttpServiceResponder} for the http request.
+   */
+  private void executeRemotely(Namespace namespace, String request, @Nullable DeltaConfig config,
+                               Class<? extends RemoteAssessmentTaskBase> remoteExecutionTaskClass,
+                               HttpServiceResponder responder)  {
+
+    DeltaConfig deltaConfig = getMacroEvaluator().evaluateMacros(namespace, config);
+    RemoteAssessmentRequest remoteRequest = new RemoteAssessmentRequest(namespace.getName(), request, deltaConfig);
+    RunnableTaskRequest runnableTaskRequest =
+      RunnableTaskRequest.getBuilder(remoteExecutionTaskClass.getName()).
+        withParam(GSON.toJson(remoteRequest)).
+        build();
+    try {
+      byte[] bytes = getContext().runTask(runnableTaskRequest);
+      responder.sendString(new String(bytes, StandardCharsets.UTF_8));
+    } catch (RemoteExecutionException e) {
+      LOG.error("Error in executing remote task "+remoteExecutionTaskClass, e);
+      RemoteTaskException remoteTaskException = e.getCause();
+      responder.sendError(
+        getExceptionCode(remoteTaskException.getRemoteExceptionClassName()), remoteTaskException.getMessage());
+    } catch (Exception e) {
+      LOG.error("Error in executing remote task "+remoteExecutionTaskClass, e);
+      responder.sendError(HttpURLConnection.HTTP_INTERNAL_ERROR, e.getMessage());
+    }
+  }
+
+  private int getExceptionCode(String exceptionClass) {
+    if (IllegalArgumentException.class.getName().equals(exceptionClass)
+      || JsonSyntaxException.class.getName().equals(exceptionClass)) {
+      return HttpURLConnection.HTTP_BAD_REQUEST;
+    }
+    return HttpURLConnection.HTTP_INTERNAL_ERROR;
   }
 }
