+++ delta-app/src/main/java/io/cdap/delta/macros/PropertyEvaluator.java
@@ -1,5 +1,5 @@
 /*
- * Copyright © 2020 Cask Data, Inc.
+ * Copyright © 2022 Cask Data, Inc.
  *
  * Licensed under the Apache License, Version 2.0 (the "License"); you may not
  * use this file except in compliance with the License. You may obtain a copy of
@@ -14,7 +14,7 @@
  * the License.
  */
 
-package io.cdap.delta.store;
+package io.cdap.delta.macros;
 
 import java.util.Map;
