+++ delta-app/src/main/java/io/cdap/delta/store/DraftService.java
@@ -16,59 +16,25 @@
 
 package io.cdap.delta.store;
 
-import io.cdap.cdap.api.plugin.InvalidPluginConfigException;
-import io.cdap.cdap.api.plugin.PluginProperties;
 import io.cdap.cdap.spi.data.transaction.TransactionRunner;
 import io.cdap.cdap.spi.data.transaction.TransactionRunners;
 import io.cdap.delta.api.Configurer;
-import io.cdap.delta.api.DeltaSource;
-import io.cdap.delta.api.SourceColumn;
-import io.cdap.delta.api.SourceTable;
-import io.cdap.delta.api.assessment.Assessment;
-import io.cdap.delta.api.assessment.ColumnAssessment;
-import io.cdap.delta.api.assessment.ColumnDetail;
-import io.cdap.delta.api.assessment.ColumnSuggestion;
-import io.cdap.delta.api.assessment.ColumnSupport;
 import io.cdap.delta.api.assessment.PipelineAssessment;
-import io.cdap.delta.api.assessment.Problem;
-import io.cdap.delta.api.assessment.StandardizedTableDetail;
-import io.cdap.delta.api.assessment.TableAssessment;
-import io.cdap.delta.api.assessment.TableAssessor;
-import io.cdap.delta.api.assessment.TableAssessorSupplier;
 import io.cdap.delta.api.assessment.TableDetail;
 import io.cdap.delta.api.assessment.TableList;
 import io.cdap.delta.api.assessment.TableNotFoundException;
-import io.cdap.delta.api.assessment.TableRegistry;
-import io.cdap.delta.api.assessment.TableSummaryAssessment;
+import io.cdap.delta.app.service.Assessor;
+import io.cdap.delta.macros.MacroEvaluator;
 import io.cdap.delta.proto.DBTable;
 import io.cdap.delta.proto.DeltaConfig;
 import io.cdap.delta.proto.DraftRequest;
-import io.cdap.delta.proto.FullColumnAssessment;
-import io.cdap.delta.proto.Plugin;
-import io.cdap.delta.proto.Stage;
 import io.cdap.delta.proto.TableAssessmentResponse;
-import io.cdap.delta.proto.TableTransformation;
-import io.cdap.transformation.ColumnRenameInfo;
-import io.cdap.transformation.DefaultMutableRowSchema;
-import io.cdap.transformation.DefaultRenameInfo;
-import io.cdap.transformation.TransformationException;
-import io.cdap.transformation.TransformationUtil;
-import io.cdap.transformation.api.Transformation;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
 import java.io.IOException;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.HashSet;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
-import java.util.Set;
-import java.util.UUID;
-import java.util.stream.Collectors;
-import javax.annotation.Nullable;
 
 /**
  * Handles logic around storage, retrieval, and assessment of drafts.
@@ -76,11 +42,13 @@ import javax.annotation.Nullable;
 public class DraftService {
   private static final Logger LOG = LoggerFactory.getLogger(DraftService.class);
   private final TransactionRunner txRunner;
-  private final PropertyEvaluator propertyEvaluator;
+  private final MacroEvaluator macroEvaluator;
+  private final Assessor pluginAssesor;
 
-  public DraftService(TransactionRunner transactionRunner, PropertyEvaluator propertyEvaluator) {
+  public DraftService(TransactionRunner transactionRunner, MacroEvaluator macroEvaluator) {
     this.txRunner = transactionRunner;
-    this.propertyEvaluator = propertyEvaluator;
+    this.macroEvaluator = macroEvaluator;
+    this.pluginAssesor = new Assessor();
   }
 
   /**
@@ -161,9 +129,30 @@ public class DraftService {
    */
   public TableList listDraftTables(DraftId draftId, Configurer configurer) throws Exception {
     Draft draft = getDraft(draftId);
-    try (TableRegistry tableRegistry = createTableRegistry(draftId.getNamespace(), draft.getConfig(), configurer)) {
-      return tableRegistry.listTables();
-    }
+    return listDraftTables(draftId, draft, configurer);
+  }
+
+  /**
+   * List the database tables readable by the source in the given draft id. An instance of the plugin will be
+   * instantiated in order to generate this list. This is an expensive operation.
+   *
+   * If the plugin cannot be instantiated, or if the table list cannot be generated due to missing information in
+   * the draft, an {@link InvalidDraftException} will be thrown.
+   *
+   * @param draftId id of the draft
+   * @param draft draft
+   * @param configurer configurer used to instantiate plugins
+   * @return list of tables readable by the source in the draft
+   * @throws DraftNotFoundException if the draft does not exist
+   * @throws InvalidDraftException if the table list cannot be fetched because the draft is invalid
+   * @throws IOException if the was an IO error fetching the table list
+   * @throws Exception if there was an error creating the table registry
+   */
+  public TableList listDraftTables(DraftId draftId, Draft draft, Configurer configurer) throws Exception {
+    DeltaConfig config = draft.getConfig();
+    Namespace namespace = draftId.getNamespace();
+    config = macroEvaluator.evaluateMacros(namespace, config);
+    return pluginAssesor.listTables(namespace.getName(), config, configurer);
   }
 
   /**
@@ -175,9 +164,7 @@ public class DraftService {
    *
    * @param draftId id of the draft
    * @param configurer configurer used to instantiate plugins
-   * @param database the table database
-   * @param  schema the table schema, only required for some databases (e.g. Oracle)
-   * @param table the table name
+   * @param dbTable the dbTable
    * @return list of tables readable by the source in the draft
    * @throws DraftNotFoundException if the draft does not exist
    * @throws InvalidDraftException if the table list cannot be fetched because the draft is invalid
@@ -185,16 +172,13 @@ public class DraftService {
    * @throws IOException if the was an IO error fetching the table detail
    * @throws Exception if there was an error creating the table registry
    */
-  public TableDetail describeDraftTable(DraftId draftId, Configurer configurer, String database,
-    @Nullable String schema, String table) throws Exception {
+  public TableDetail describeDraftTable(DraftId draftId, Configurer configurer, DBTable dbTable) throws Exception {
     Draft draft = getDraft(draftId);
-    try (TableRegistry tableRegistry = createTableRegistry(draftId.getNamespace(), draft.getConfig(), configurer)) {
-      if (schema == null) {
-        return tableRegistry.describeTable(database, table);
-      } else {
-        return tableRegistry.describeTable(database, schema, table);
-      }
-    }
+    DeltaConfig config = draft.getConfig();
+    Namespace namespace = draftId.getNamespace();
+    config = macroEvaluator.evaluateMacros(namespace, config);
+
+    return pluginAssesor.describeTable(draftId.getNamespace().getName(), config, configurer, dbTable);
   }
 
   /**
@@ -204,7 +188,7 @@ public class DraftService {
    * If a plugin cannot be instantiated due to missing draft information,
    * an {@link InvalidDraftException} will be thrown.
    *
-   * @param draftId id of the draft
+   * @param draft id of the draft
    * @param dbTable the database, the schema, the table to be assessed
    * @param configurer configurer used to instantiate plugins
    * @return list of tables readable by the source in the draft
@@ -215,9 +199,9 @@ public class DraftService {
    * @throws Exception if there was an error creating the table registry
    * @throws IllegalArgumentException if the table is not selected in the draft
    */
-  public TableAssessmentResponse assessTable(DraftId draftId, Configurer configurer, DBTable dbTable) throws Exception {
-    Draft draft = getDraft(draftId);
-    return assessTable(draftId.getNamespace(), draft.getConfig(), configurer, dbTable);
+  public TableAssessmentResponse assessTable(Namespace namespace, Draft draft, Configurer configurer, DBTable dbTable) throws Exception {
+    DeltaConfig config = draft.getConfig();
+    return assessTable(namespace, config, configurer, dbTable);
   }
 
   /**
@@ -235,54 +219,16 @@ public class DraftService {
    */
   public TableAssessmentResponse assessTable(Namespace namespace, DeltaConfig deltaConfig, Configurer configurer,
                                               DBTable dbTable) throws Exception {
-    deltaConfig = evaluateMacros(namespace, deltaConfig);
-    SourceTable selectedTable = deltaConfig.getTables().stream().filter(
-      streamTable -> dbTable.getDatabase().equals(streamTable.getDatabase()) &&
-        dbTable.getTable().equals(streamTable.getTable()) &&
-        (dbTable.getSchema() == streamTable.getSchema() || dbTable.getSchema() != null
-          && dbTable.getSchema().equals(streamTable.getSchema()))).findAny()
-      .orElseThrow(() -> new IllegalArgumentException(String
-        .format("Table '%s' in database '%s' and schema '%s' is not a selected table in the draft", dbTable.getTable(),
-                dbTable.getDatabase(), dbTable.getSchema())));
-    Stage target = deltaConfig.getTarget();
-    if (target == null) {
-      throw new InvalidDraftException("Cannot assess a table without a configured target.");
-    }
-
-    Map<String, TableTransformation> tableLevelTransformations =
-      TransformationUtil.getTableLevelTransformations(deltaConfig);
-    Map<String, List<Transformation>> transformations = new HashMap<>();
-    try {
-      transformations = TransformationUtil.loadTransformations(configurer, tableLevelTransformations, selectedTable);
-    } catch (TransformationException e) {
-      LOG.debug("Failed to apply transformation: ", e);
-      List<Problem> transformationErrors = new ArrayList<>();
-      transformationErrors.add(new Problem("Transformation Loading failed", e.getMessage(),
-                               "Please ensure the applied transformation's plugin is uploaded'",
-                               "", Problem.Severity.ERROR, e.getTableName(), e.getColumnName()));
-
-      return new TableAssessmentResponse(Collections.emptyList(), Collections.emptyList(), transformationErrors);
-    }
-
-    try (TableRegistry tableRegistry = createTableRegistry(namespace, deltaConfig, configurer);
-         TableAssessor<TableDetail> sourceTableAssessor = createTableAssessor(configurer, deltaConfig.getSource(),
-           deltaConfig.getTables());
-         TableAssessor<StandardizedTableDetail> targetTableAssessor = createTableAssessor(configurer, target,
-           deltaConfig.getTables())) {
-
-      TableAssessmentResponse assessment = assessTable(selectedTable, tableRegistry, sourceTableAssessor,
-                                                       targetTableAssessor, transformations);
-      return assessment;
-    }
+    deltaConfig = macroEvaluator.evaluateMacros(namespace, deltaConfig);
+    return pluginAssesor.assessTable(namespace.getName(), deltaConfig, configurer, dbTable);
   }
 
   /**
    * Assess the entire pipeline based on the tables the source will read and the capabilities of the target.
    * This will fetch draft from the store.
    */
-  public PipelineAssessment assessPipeline(DraftId draftId, Configurer configurer) throws Exception {
-    Draft draft = getDraft(draftId);
-    return assessPipeline(draftId.getNamespace(), draft.getConfig(), configurer);
+  public PipelineAssessment assessPipeline(Namespace namespace, Draft draft, Configurer configurer) throws Exception {
+    return assessPipeline(namespace, draft.getConfig(), configurer);
   }
 
   /**
@@ -304,306 +250,7 @@ public class DraftService {
    */
   public PipelineAssessment assessPipeline(Namespace namespace, DeltaConfig deltaConfig, Configurer configurer)
     throws Exception {
-    deltaConfig.validatePipeline();
-    deltaConfig = evaluateMacros(namespace, deltaConfig);
-
-    try (TableRegistry tableRegistry = createTableRegistry(namespace, deltaConfig, configurer);
-         TableAssessor<TableDetail> sourceTableAssessor = createTableAssessor(configurer, deltaConfig.getSource(),
-           deltaConfig.getTables());
-         TableAssessor<StandardizedTableDetail> targetTableAssessor =
-           createTableAssessor(configurer, deltaConfig.getTarget(), deltaConfig.getTables())) {
-
-      List<Problem> missingFeatures = new ArrayList<>();
-      List<Problem> connectivityIssues = new ArrayList<>();
-      List<Problem> transformationIssues = new ArrayList<>();
-      List<TableSummaryAssessment> tableAssessments = new ArrayList<>();
-
-      Assessment sourceAssessment = sourceTableAssessor.assess();
-      Assessment targetAssessment = targetTableAssessor.assess();
-      missingFeatures.addAll(sourceAssessment.getFeatures());
-      missingFeatures.addAll(targetAssessment.getFeatures());
-      connectivityIssues.addAll(sourceAssessment.getConnectivity());
-      connectivityIssues.addAll(targetAssessment.getConnectivity());
-
-      // if no source tables are given, this means all tables should be read
-      List<SourceTable> tablesToAssess = deltaConfig.getTables();
-      if (tablesToAssess.isEmpty()) {
-        tablesToAssess = tableRegistry.listTables().getTables().stream()
-          .map(t -> new SourceTable(t.getDatabase(), t.getTable()))
-          .collect(Collectors.toList());
-      }
-
-      Map<String, TableTransformation> tableLevelTransformations =
-        TransformationUtil.getTableLevelTransformations(deltaConfig);
-      // go through all tables that the pipeline should read, fetching detail about each of the tables
-      for (SourceTable sourceTable : tablesToAssess) {
-        String db = sourceTable.getDatabase();
-        String table = sourceTable.getTable();
-        try {
-          Map<String, List<Transformation>> transformations = TransformationUtil.loadTransformations(configurer,
-                                                                                        tableLevelTransformations,
-                                                                                        sourceTable);
-          TableAssessmentResponse assessment = assessTable(sourceTable, tableRegistry, sourceTableAssessor,
-                                                           targetTableAssessor, transformations);
-          missingFeatures.addAll(assessment.getFeatures());
-          tableAssessments.add(summarize(db, sourceTable.getSchema(), sourceTable, assessment));
-          transformationIssues.addAll(assessment.getTransformationIssues());
-        } catch (TableNotFoundException e) {
-          connectivityIssues.add(
-            new Problem("Table Not Found",
-                        String.format("Table '%s' in database '%s' was not found.", table, db),
-                        "Check the table information and permissions",
-                        null));
-        } catch (IOException e) {
-          connectivityIssues.add(
-            new Problem("Table Describe Error",
-                        String.format("Unable to fetch details about table '%s' in database '%s': %s",
-                                      table, db, e.getMessage()),
-                        "Check permissions and database connectivity",
-                        null));
-        } catch (TransformationException e) {
-          LOG.debug("Failed to apply transformation: ", e);
-          transformationIssues.add(new Problem("Transformation Loading failed", e.getMessage(),
-                                      "Please ensure the applied transformation's plugin is uploaded'",
-                                      "", Problem.Severity.ERROR, e.getTableName(), e.getColumnName()));
-        }
-      }
-      return new PipelineAssessment(tableAssessments, missingFeatures, connectivityIssues, transformationIssues);
-    }
-  }
-
-  private TableAssessmentResponse assessTable(SourceTable sourceTable, TableRegistry tableRegistry,
-                                              TableAssessor<TableDetail> sourceTableAssessor,
-                                              TableAssessor<StandardizedTableDetail> targetTableAssesor,
-                                              Map<String, List<Transformation>> transformations)
-    throws IOException, TableNotFoundException {
-    String db = sourceTable.getDatabase();
-    String table = sourceTable.getTable();
-    String schema = sourceTable.getSchema();
-    Set<String> columnWhitelist = sourceTable.getColumns().stream()
-      .map(SourceColumn::getName)
-      .collect(Collectors.toSet());
-
-    // fetch detail about the table, then filter out columns that will not be read by the source
-    TableDetail detail;
-    if (schema == null) {
-      detail = tableRegistry.describeTable(db, table);
-    } else {
-      detail = tableRegistry.describeTable(db, schema, table);
-    }
-
-    List<Problem> missingFeatures = new ArrayList<>(detail.getFeatures());
-    List<String> unselectedPrimaryKeys = detail.getPrimaryKey().stream()
-      // if there are no columns specified, it means all columns of the table are selected
-      .filter(columnWhitelist.isEmpty() ? col -> false : col -> !columnWhitelist.contains(col))
-      .collect(Collectors.toList());
-    if (unselectedPrimaryKeys.size() == 1) {
-      missingFeatures.add(
-        new Problem("Missing Primary Key",
-                    String.format("Column '%s' is part of the primary key for table '%s' in database '%s', " +
-                                    "but is not selected to be replicated", unselectedPrimaryKeys.get(0),
-                                  table, db),
-                    "Please make sure this column been selected",
-                    "This can result in different data at the target than at the source"));
-    } else if (unselectedPrimaryKeys.size() > 1) {
-      missingFeatures.add(
-        new Problem("Missing Primary Key",
-                    String.format("Columns '%s' are part of the primary key for table '%s' in database '%s', " +
-                                    "but are not selected to be replicated",
-                                  String.join(",", unselectedPrimaryKeys),
-                                  table, db),
-                    "Please make sure columns been selected",
-                    "This can result in different data at the target than at the source"));
-    }
-
-    List<ColumnDetail> selectedColumns = detail.getColumns().stream()
-      // if there are no columns specified, it means all columns should be read
-      .filter(columnWhitelist.isEmpty() ? col -> true : col -> columnWhitelist.contains(col.getName()))
-      .collect(Collectors.toList());
-    TableDetail filteredDetail = TableDetail.builder(db, table, schema)
-      .setPrimaryKey(detail.getPrimaryKey())
-      .setColumns(selectedColumns)
-      .setFeatures(missingFeatures)
-      .build();
-    TableAssessment srcAssessment = sourceTableAssessor.assess(filteredDetail);
-
-    StandardizedTableDetail standardizedDetail = tableRegistry.standardize(filteredDetail);
-
-    ColumnRenameInfo columnRenameInfo = new DefaultRenameInfo(Collections.emptyMap());
-    List<Problem> transformationIssues = new ArrayList<>();
-    if (!transformations.isEmpty()) {
-      try {
-        DefaultMutableRowSchema rowSchema = TransformationUtil.transformSchema(table, standardizedDetail.getSchema(),
-                                                                               transformations);
-       standardizedDetail = new StandardizedTableDetail(standardizedDetail.getDatabase(),
-                                                        standardizedDetail.getTable(),
-                                                        standardizedDetail.getPrimaryKey(),
-                                                        rowSchema.toSchema());
-       columnRenameInfo = rowSchema.getRenameInfo();
-      } catch (TransformationException e) {
-        LOG.debug("Failed to apply transformation: ", e);
-        transformationIssues.add(new Problem("Transformation failed", e.getMessage(),
-                    "Please ensure the applied transformation is valid",
-                    "The job cannot be deployed with invalid transformations", Problem.Severity.ERROR,
-                     e.getTableName(), e.getColumnName()));
-      }
-    }
-    TableAssessment targetAssessment = targetTableAssesor.assess(standardizedDetail);
-
-    return merge(srcAssessment, targetAssessment, columnRenameInfo, transformationIssues);
-  }
-
-  /**
-   * Merge the assessment from the source and target into a single assessment.
-   * This amounts to merging the assessment for each column.
-   */
-  private TableAssessmentResponse merge(TableAssessment srcAssessment, TableAssessment targetAssessment,
-                                        ColumnRenameInfo columnRenameInfo, List<Problem> transformationIssues) {
-    Map<String, ColumnAssessment> targetColumns = targetAssessment.getColumns().stream()
-      .filter(c -> c.getSourceName() != null)
-      .collect(Collectors.toMap(ColumnAssessment::getSourceName, c -> c));
-
-    List<FullColumnAssessment> fullColumns = new ArrayList<>();
-    Set<String> addedColumns = new HashSet<>();
-    // add columns from the source
-    for (ColumnAssessment srcColumn : srcAssessment.getColumns()) {
-      String name = columnRenameInfo.getNewName(srcColumn.getName());
-      ColumnAssessment targetColumn = targetColumns.get(name);
-      if (targetColumn != null) {
-        fullColumns.add(merge(srcColumn, targetColumn));
-      } else {
-        fullColumns.add(new FullColumnAssessment(srcColumn.getSupport(), srcColumn.getName(), srcColumn.getType(),
-                                                 null, null, srcColumn.getSuggestion()));
-      }
-      addedColumns.add(name);
-    }
-    // add columns present only in the target and not the source
-    targetAssessment.getColumns().stream()
-      .filter(t -> !addedColumns.contains(t.getName()))
-      .forEach(t -> fullColumns.add(new FullColumnAssessment(t.getSupport(), null, null, t.getName(), t.getType(),
-                                                             t.getSuggestion())));
-
-    List<Problem> features = new ArrayList<>(srcAssessment.getFeatureProblems());
-    features.addAll(targetAssessment.getFeatureProblems());
-    return new TableAssessmentResponse(fullColumns, features, transformationIssues);
-  }
-
-  /**
-   * Merge the assessment form the source and target into a single assessment.
-   * If one plugin supports a column but the other does not, that column is not supported from a pipeline perspective.
-   * Similarly if one plugin fully supports a column but the other only partially supports it, that column is
-   * partially supported from a pipeline perspective.
-   * If both plugins do not support the column, or both partially support the column, the assessment from the source
-   * is taken, because that is what needs to be addressed first.
-   */
-  private FullColumnAssessment merge(ColumnAssessment srcAssessment, ColumnAssessment targetAssessment) {
-    ColumnSupport support = ColumnSupport.YES;
-    if (srcAssessment.getSupport() == ColumnSupport.NO || targetAssessment.getSupport() == ColumnSupport.NO) {
-      support = ColumnSupport.NO;
-    } else if (srcAssessment.getSupport() == ColumnSupport.PARTIAL ||
-      targetAssessment.getSupport() == ColumnSupport.PARTIAL) {
-      support = ColumnSupport.PARTIAL;
-    }
-    ColumnSuggestion suggestion = srcAssessment.getSuggestion();
-    if (suggestion == null) {
-      suggestion = targetAssessment.getSuggestion();
-    }
-    return new FullColumnAssessment(support, srcAssessment.getName(), srcAssessment.getType(),
-                                    targetAssessment.getName(), targetAssessment.getType(),
-                                    suggestion);
-  }
-
-  private TableRegistry createTableRegistry(Namespace namespace, DeltaConfig deltaConfig, Configurer configurer)
-    throws Exception {
-    deltaConfig = evaluateMacros(namespace, deltaConfig);
-    Stage stage = deltaConfig.getSource();
-
-    Plugin pluginConfig = stage.getPlugin();
-    DeltaSource deltaSource;
-    try {
-      deltaSource = configurer.usePlugin(pluginConfig.getType(), pluginConfig.getName(), UUID.randomUUID().toString(),
-                                         PluginProperties.builder().addAll(pluginConfig.getProperties()).build());
-    } catch (InvalidPluginConfigException e) {
-      throw new InvalidDraftException(String.format("Unable to instantiate source plugin: %s", e.getMessage()), e);
-    }
-
-    if (deltaSource == null) {
-      throw new InvalidDraftException(String.format("Unable to find plugin '%s'", pluginConfig.getName()));
-    }
-
-    return deltaSource.createTableRegistry(configurer);
-  }
-
-  private <T> TableAssessor<T> createTableAssessor(Configurer configurer, Stage stage, List<SourceTable> tables) {
-    Plugin pluginConfig = stage.getPlugin();
-    TableAssessorSupplier<T> plugin;
-    try {
-      plugin = configurer.usePlugin(pluginConfig.getType(), pluginConfig.getName(), UUID.randomUUID().toString(),
-                                    PluginProperties.builder().addAll(pluginConfig.getProperties()).build());
-    } catch (InvalidPluginConfigException e) {
-      throw new InvalidDraftException(
-        String.format("Unable to instantiate plugin for stage '%s': %s", stage.getName(), e.getMessage()), e);
-    }
-
-    if (plugin == null) {
-      throw new InvalidDraftException(String.format("Unable to find plugin '%s' for stage '%s'",
-                                                    pluginConfig.getName(), stage.getName()));
-    }
-
-    try {
-      return plugin.createTableAssessor(configurer, tables);
-    } catch (Exception e) {
-      throw new InvalidDraftException(String.format("Unable to instantiate table assessor for stage '%s': %s",
-                                                    stage.getName(), e.getMessage()), e);
-    }
-  }
-
-  private TableSummaryAssessment summarize(String db, @Nullable String schema, SourceTable table,
-    TableAssessmentResponse assessment) {
-    int numUnsupported = 0;
-    int numPartial = 0;
-    int numColumns = 0;
-    for (FullColumnAssessment columnAssessment : assessment.getColumns()) {
-      if (columnAssessment.getSupport() == ColumnSupport.NO) {
-        numUnsupported++;
-      } else if (columnAssessment.getSupport() == ColumnSupport.PARTIAL) {
-        Set<SourceColumn> columns = table.getColumns();
-        SourceColumn existingColumn = columns.stream()
-          .filter(column -> columnAssessment.getSourceName().equals(column.getName()))
-          .findAny()
-          .orElse(null);
-
-        if (existingColumn == null || !existingColumn.isSuppressWarning()) {
-          numPartial++;
-        }
-      }
-      numColumns++;
-    }
-    return new TableSummaryAssessment(db, table.getTable(), numColumns, numUnsupported, numPartial, schema);
-  }
-
-  private DeltaConfig evaluateMacros(Namespace namespace, DeltaConfig config) {
-    DeltaConfig.Builder builder = DeltaConfig.builder()
-      .setDescription(config.getDescription())
-      .setResources(config.getResources())
-      .setSource(config.getSource())
-      .setOffsetBasePath(config.getOffsetBasePath())
-      .setTables(config.getTables())
-      .setTableTransformations(config.getTableTransformations())
-      .setSource(evaluateMacros(namespace.getName(), config.getSource()));
-
-    Stage target = config.getTarget();
-    if (target != null) {
-      builder.setTarget(evaluateMacros(namespace.getName(), target));
-    }
-
-    return builder.build();
-  }
-
-  private Stage evaluateMacros(String namespace, Stage stage) {
-    Plugin plugin = stage.getPlugin();
-    Map<String, String> evaluatedProperties = propertyEvaluator.evaluate(namespace, plugin.getProperties());
-    Plugin evaluatedPlugin = new Plugin(plugin.getName(), plugin.getType(), evaluatedProperties, plugin.getArtifact());
-    return new Stage(stage.getName(), evaluatedPlugin);
+    deltaConfig = macroEvaluator.evaluateMacros(namespace, deltaConfig);
+    return pluginAssesor.assessPipeline(namespace.getName(), deltaConfig, configurer);
   }
 }
